Logo, transitions
Added a logo that matches the color scheme of the website and added
transitions. The transitions aren't done, but they work relatively well.

Media queries for the logo are done, and the logo should look good, and
not obstruct anything no matter the viewing resolution.

On super low resolutions, I just hide the logo.